use jwt in analytics request
