+++ cloud/client.go
@@ -89,6 +89,7 @@ type Client interface {
 	DeleteCachedToken() error
 	DisableSSHKeyGuessing()
 	SetAuthTokenDir(path string)
+	SendAnalytics(data *EarthlyAnalytics) error
 }
 
 type request struct {
@@ -967,6 +968,42 @@ func (c *client) WhoAmI() (string, string, bool, error) {
 	}
 
 	return email, authType, writeAccess, nil
+}
+
+type EarthlyAnalytics struct {
+	Key              string                    `json:"key"`
+	InstallID        string                    `json:"install_id"`
+	Version          string                    `json:"version"`
+	Platform         string                    `json:"platform"`
+	GitSHA           string                    `json:"git_sha"`
+	ExitCode         int                       `json:"exit_code"`
+	CI               string                    `json:"ci_name"`
+	RepoHash         string                    `json:"repo_hash"`
+	ExecutionSeconds float64                   `json:"execution_seconds"`
+	Terminal         bool                      `json:"terminal"`
+	Counts           map[string]map[string]int `json:"counts"`
+}
+
+func (c *client) SendAnalytics(data *EarthlyAnalytics) error {
+	payload, err := json.Marshal(data)
+	if err != nil {
+		return errors.Wrap(err, "failed to marshal data")
+	}
+	opts := []requestOpt{
+		withBody(string(payload)),
+		withHeader("Content-Type", "application/json; charset=utf-8"),
+	}
+	if c.authToken != "" {
+		opts = append(opts, withAuth())
+	}
+	status, _, err := c.doCall("PUT", "/analytics", opts...)
+	if err != nil {
+		return errors.Wrap(err, "failed sending analytics")
+	}
+	if status != http.StatusAccepted {
+		return errors.Errorf("unexpected response from analytics server: %d", status)
+	}
+	return nil
 }
 
 func (c *client) migrateOldToken() error {
